Updated example to use the latest multer :octocat: #75
