+++ example/package.json
@@ -13,7 +13,7 @@
   "dependencies": {
     "body-parser": "^1.15.0",
     "express": "^4.13.4",
-    "multer": "^0.1.8",
+    "multer": "^1.1.0",
     "react": "^0.14.7",
     "react-dom": "^0.14.7",
     "react-dropzone-component": "github:felixrieseberg/react-dropzone-component"
