+++ example/server.js
@@ -7,18 +7,8 @@ var express     = require('express'),
 
 app.set('port', process.env.PORT || 3000);
 app.use(express.static('../'));
-app.use(new MulterImpl({}).init());
+require('./src-server/multerImpl')(app); //Call the multerImpl and pass in app state to it
 
-app.post('/uploadHandler', function (req, res) {
-    // This block is only relevant to users
-    // interested in custom parameters - you
-    // can delete/ignore it as you wish
-    if (req.body) {
-        console.dir(req.body);
-    }
-
-    res.sendStatus(200);
-});
 
 module.exports = app.listen(app.get('port'), function() {
     console.log('Express server listening on port ' + app.get('port'));
