--- example2/index.html
@@ -1,20 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-        <meta http-equiv='Content-type' content='text/html; charset=utf-8'>
-        <title>React Dropzone Example</title>
-        <link rel="stylesheet" type="text/css" href="./styles/example.css" />
-        <link rel="stylesheet" type="text/css" href="../styles/filepicker.css" />
-        <link rel="stylesheet" type="text/css" href="../node_modules/dropzone/dist/min/dropzone.min.css" />
-    </head>
-    <body>
-        <div class="intro">
-        <h2><a href="https://github.com/felixrieseberg/React-Filepicker">React Filepicker Example</a></h2>
-        <p>This is a simple example for React-Dropzone. To enable support for uploads, make sure to run this example by running <pre>npm start</pre> in the "example" folder and visiting <a href="http://localhost:3000/example/">http://localhost:3000/example/</a>.</p>
-        </div>
-        <div class="example">
-            <div id="content"></div>
-        </div>
-        <script src="built/bundle.js"></script>
-    </body>
-</html>
