--- example2/src/example_params.js
@@ -1,25 +1,0 @@
-'use strict';
-
-var React             = require('react');
-var DropzoneComponent = require('react-dropzone-component');
-
-// Add params to your dropzonejs configuration object, like so:
-var djsConfig = {
-    addRemoveLinks: true,
-    params: {
-        myParam: 'Hello from a parameter!',
-        anotherParam: 43
-    }
-};
-
-var componentConfig = {
-    showFiletypeIcon: false,
-    postUrl: '/uploadHandler'
-};
-
-// Render
-React.render(React.createElement(DropzoneComponent, { 
-        config: componentConfig,
-        djsConfig: djsConfig 
-    }), document.getElementById('content')
-);
