--- example2/webpack.config.js
@@ -1,6 +1,0 @@
-module.exports = {
-    entry: "./src/example_default.js",
-    output: {
-        filename: "./built/bundle.js"
-    }
-};
